refactor(messageThunk): extract DB operations to messageDB.ts
Extract 10 database operation functions from the monolithic messageThunk.ts
(2,255 lines) into a dedicated messageDB.ts module:

- getRawTopic
- updateFileCount
- deleteMessageFromDB
- deleteMessagesFromDB
- clearMessagesFromDB
- saveMessageAndBlocksToDB
- updateMessage
- updateSingleBlock
- bulkAddBlocks
- updateBlocks

This reduces messageThunk.ts by ~147 lines and separates database
operations from the stream/thunk logic for better maintainability.
